create a utility to set coordinate precision, re #5476
